提交

+++ project/chameleon/chameleon.h
@@ -29,7 +29,7 @@ public:
     chameleon(QWidget *parent = nullptr);
     ~chameleon();
     void initWindow();
-    void initButton(bool a,int b);
+    void initButton(bool a,double b);
     void Imageset(QLabel* image,QPixmap pixmap);
     void initLayout();  //初始化布局管理器  这个主要用于加载角色，放在布局管理器中角色会随着窗口的伸缩变换
